fix: 修复编辑状态下有查询组件的删除之后图表还是维持过滤状态问题
